refactor: move getGitIconName and reuse GitProviderIcon
